feature: hour type
